requirements.txt
Flask
gunicorn
python-dotenv
pymongo
boto3
PyJWT
requests
flask-swagger-ui
pytest
fastapi
uvicorn
motor
pydantic-settings
httpx
python-multipart
cryptography